Split @ydbjs/drizzle-adapter into ., /schema, /sql, /migrator entries
The root used to dump 90+ exports — driver, columns, DDL builders, YQL
helpers, relations — into a single barrel. Splits them into four subpath
entries with sharply defined surfaces:

- `.` keeps only the bootstrap surface (createDrizzle, drizzle, YdbDriver,
  the error classes, relations re-export)
- `/schema` owns table declaration: ydbTable, columns, primaryKey, unique,
  indexes, table options, schema type helpers
- `/sql` owns YQL expression helpers: builtins, digests, knn, set-operators,
  grouping, windows, pragmas, scripts, dialect config types
- `/migrator` owns migrate() plus the build*Sql DDL families and their types

Surface tests now live next to each entry file and assert the exact name set;
check-package-surface.mjs verifies each entrypoint imports cleanly and that
schema/sql/migrator symbols don't leak across each other. Internal tests and
the example switched to the new subpath imports.
