+++ third-parties/drizzle-adapter/src/index.ts
@@ -1,18 +1,14 @@
-export type {
-	YdbDeleteConfig,
-	YdbDialectMigration,
-	YdbDialectMigrationConfig,
-	YdbFlatRelationalQueryConfig,
-	YdbInsertConfig,
-	YdbJoinConfig,
-	YdbJoinType,
-	YdbRelationalQueryConfig,
-	YdbRelationalQueryResult,
-	YdbSelectConfig,
-	YdbSetOperatorConfig,
-	YdbSetOperatorSource,
-	YdbUpdateConfig,
-} from './ydb/dialect.js'
+// Bootstrap surface: how you connect, what `drizzle()` returns, and the error
+// classes you catch. For schemas, YQL expressions, and migrations import from
+// `@ydbjs/drizzle-adapter/schema`, `/sql`, and `/migrator` respectively.
+
+export {
+	createDrizzle,
+	drizzle,
+	type YdbDrizzleConfig,
+	type YdbDrizzleDatabase,
+	type YdbDrizzleOptions,
+} from './ydb/createDrizzle.js'
 export {
 	YdbDriver,
 	type YdbDriverOptions,
@@ -37,230 +33,7 @@ export {
 	type YdbQueryErrorKind,
 } from './ydb/errors.js'
 export type { YdbTransactionScope } from './ydb-core/db.js'
-export {
-	type YdbSchemaDefinition,
-	type YdbSchemaRelations,
-	type YdbSchemaWithoutTables,
-} from './ydb-core/schema.types.js'
-export {
-	asTable,
-	commit,
-	crc32c,
-	crc64,
-	cube,
-	currentUtcDate,
-	currentUtcDatetime,
-	currentUtcTimestamp,
-	declareParam,
-	defineAction,
-	distinctHint,
-	doAction,
-	doBlock,
-	except,
-	groupKey,
-	grouping,
-	groupingSets,
-	hop,
-	hopEnd,
-	hopStart,
-	intersect,
-	intoResult,
-	kMeansTreeSearchTopSize,
-	knnCosineDistance,
-	knnCosineSimilarity,
-	knnDistance,
-	knnEuclideanDistance,
-	knnInnerProductSimilarity,
-	knnManhattanDistance,
-	knnSimilarity,
-	matchRecognize,
-	maxOf,
-	minOf,
-	numericHash,
-	pragma,
-	random,
-	randomNumber,
-	randomUuid,
-	rollup,
-	sessionStart,
-	sessionWindow,
-	uniqueHint,
-	union,
-	unionAll,
-	unwrap,
-	values,
-	valuesTable,
-	windowDefinition,
-	xxHash,
-	yqlScript,
-	type YdbActionParameter,
-	type YdbFlattenConfig,
-	type YdbFlattenMode,
-	type YdbGroupingSet,
-	type YdbKnnDistanceFunction,
-	type YdbKnnSimilarityFunction,
-	type YdbMatchRecognizeConfig,
-	type YdbSampleConfig,
-	type YdbScriptExpression,
-	type YdbScriptPrimitive,
-	type YdbUniqueDistinctHint,
-	type YdbValuesOptions,
-	type YdbValuesPrimitive,
-	type YdbValuesRow,
-	type YdbWindowClause,
-	type YdbWindowDefinitionConfig,
-} from './ydb-core/query-builders/select.js'
-export {
-	index,
-	indexView,
-	uniqueIndex,
-	vectorIndexView,
-	vectorIndex,
-	type YdbVectorDistance,
-	type YdbVectorKMeansTreeOptions,
-	type YdbVectorSimilarity,
-	type YdbVectorType,
-} from './ydb-core/indexes.js'
-export { migrate, type YdbMigrateConfig, type YdbMigratorConfig } from './ydb/migrator.js'
-export {
-	buildAddColumnsSql,
-	buildAddChangefeedSql,
-	buildAddColumnFamilySql,
-	buildAddIndexSql,
-	buildAlterAsyncReplicationSql,
-	buildAlterGroupSql,
-	buildAlterTableSql,
-	buildAlterColumnFamilySql,
-	buildAlterColumnSetFamilySql,
-	buildAlterTableResetOptionsSql,
-	buildAlterTableSetOptionsSql,
-	buildAlterTopicSql,
-	buildAlterTransferSql,
-	buildAlterUserSql,
-	buildAnalyzeSql,
-	buildCreateAsyncReplicationSql,
-	buildCreateGroupSql,
-	buildCreateSecretSql,
-	buildCreateTableSql,
-	buildCreateTopicSql,
-	buildCreateTransferSql,
-	buildCreateUserSql,
-	buildCreateViewSql,
-	buildDropAsyncReplicationSql,
-	buildDropColumnsSql,
-	buildDropChangefeedSql,
-	buildDropGroupSql,
-	buildDropIndexSql,
-	buildDropTableSql,
-	buildDropTopicSql,
-	buildDropTransferSql,
-	buildDropUserSql,
-	buildDropViewSql,
-	buildGrantSql,
-	buildMigrationLockTableBootstrapSql,
-	buildMigrationSql,
-	buildRevokeSql,
-	buildRenameTableSql,
-	buildShowCreateSql,
-	type YdbAccessPermission,
-	type YdbAccessPermissions,
-	type YdbAlterAsyncReplicationOperation,
-	type YdbAlterAsyncReplicationOptions,
-	type YdbAlterGroupOperation,
-	type YdbAlterTableAction,
-	type YdbAlterTableOperation,
-	type YdbAlterTopicAction,
-	type YdbAlterTopicOperation,
-	type YdbAlterTransferOperation,
-	type YdbAlterTransferOptions,
-	type YdbAlterUserOperation,
-	type YdbAnalyzeOperation,
-	type YdbAsyncReplicationConsistencyLevel,
-	type YdbAsyncReplicationOptions,
-	type YdbAsyncReplicationTarget,
-	type YdbChangefeedFormat,
-	type YdbChangefeedMode,
-	type YdbChangefeedOptions,
-	type YdbCreateAsyncReplicationOperation,
-	type YdbCreateGroupOperation,
-	type YdbCreateSecretOperation,
-	type YdbCreateTopicOptions,
-	type YdbCreateTopicOperation,
-	type YdbCreateTransferOperation,
-	type YdbCreateUserOperation,
-	type YdbCreateViewOperation,
-	type YdbCreateViewOptions,
-	type YdbDropAsyncReplicationOperation,
-	type YdbDropChangefeedOperation,
-	type YdbDropGroupOperation,
-	type YdbDropTopicOperation,
-	type YdbDropTransferOperation,
-	type YdbDropUserOperation,
-	type YdbDropViewOperation,
-	type YdbGrantOperation,
-	type YdbRenameTableOperation,
-	type YdbRevokeOperation,
-	type YdbShowCreateObjectType,
-	type YdbShowCreateOperation,
-	type YdbTopicConsumer,
-	type YdbTransferOptions,
-	type YdbUserOptions,
-	type YdbMigrationLockConfig,
-	type YdbInlineMigration,
-	type YdbMigrationOperation,
-	type YdbMigrationRecoveryConfig,
-	type YdbMigrationStatus,
-} from './ydb/migration-ddl.js'
-export { primaryKey } from './ydb-core/primary-keys.js'
-export { unique } from './ydb-core/unique-constraint.js'
-export {
-	createDrizzle,
-	drizzle,
-	type YdbDrizzleConfig,
-	type YdbDrizzleDatabase,
-	type YdbDrizzleOptions,
-} from './ydb/createDrizzle.js'
+// Relations are re-exported so schemas declared with
+// `@ydbjs/drizzle-adapter/schema` can stand up the relations API from a single
+// install without depending on `drizzle-orm` directly.
 export { createMany as many, createOne as one, relations } from 'drizzle-orm'
-export { ydbTable, ydbTableCreator, type YdbTable, type YdbTableFn } from './ydb-core/table.js'
-export {
-	columnFamily,
-	partitionByHash,
-	rawTableOption,
-	tableOptions,
-	ttl,
-	type YdbColumnFamilyOptions,
-	type YdbTableOptionValue,
-	type YdbTtlAction,
-	type YdbTtlUnit,
-} from './ydb-core/table-options.js'
-export { customType } from './ydb-core/columns/custom.js'
-export { integer, int } from './ydb-core/columns/integer.js'
-export { text } from './ydb-core/columns/text.js'
-export {
-	bigint,
-	binary,
-	boolean,
-	bytes,
-	date,
-	date32,
-	datetime,
-	datetime64,
-	decimal,
-	double,
-	dyNumber,
-	float,
-	int8,
-	int16,
-	interval,
-	interval64,
-	json,
-	jsonDocument,
-	timestamp,
-	timestamp64,
-	uint8,
-	uint16,
-	uint32,
-	uint64,
-	uuid,
-	yson,
-} from './ydb-core/columns/types.js'
