+++ third-parties/drizzle-adapter/src/migrator.ts
@@ -1,0 +1,96 @@
+// Migration plumbing: the `migrate()` entry point plus the DDL builders the
+// dialect uses to emit YQL for schema changes. Most apps only need `migrate`;
+// the `build*Sql` functions are exposed for tooling that wants to render or
+// inspect the same statements without running them.
+
+export { migrate, type YdbMigrateConfig, type YdbMigratorConfig } from './ydb/migrator.js'
+export type { YdbDialectMigration, YdbDialectMigrationConfig } from './ydb/dialect.js'
+export {
+	buildAddChangefeedSql,
+	buildAddColumnFamilySql,
+	buildAddColumnsSql,
+	buildAddIndexSql,
+	buildAlterAsyncReplicationSql,
+	buildAlterColumnFamilySql,
+	buildAlterColumnSetFamilySql,
+	buildAlterGroupSql,
+	buildAlterTableResetOptionsSql,
+	buildAlterTableSetOptionsSql,
+	buildAlterTableSql,
+	buildAlterTopicSql,
+	buildAlterTransferSql,
+	buildAlterUserSql,
+	buildAnalyzeSql,
+	buildCreateAsyncReplicationSql,
+	buildCreateGroupSql,
+	buildCreateSecretSql,
+	buildCreateTableSql,
+	buildCreateTopicSql,
+	buildCreateTransferSql,
+	buildCreateUserSql,
+	buildCreateViewSql,
+	buildDropAsyncReplicationSql,
+	buildDropChangefeedSql,
+	buildDropColumnsSql,
+	buildDropGroupSql,
+	buildDropIndexSql,
+	buildDropTableSql,
+	buildDropTopicSql,
+	buildDropTransferSql,
+	buildDropUserSql,
+	buildDropViewSql,
+	buildGrantSql,
+	buildMigrationLockTableBootstrapSql,
+	buildMigrationSql,
+	buildRenameTableSql,
+	buildRevokeSql,
+	buildShowCreateSql,
+	type YdbAccessPermission,
+	type YdbAccessPermissions,
+	type YdbAlterAsyncReplicationOperation,
+	type YdbAlterAsyncReplicationOptions,
+	type YdbAlterGroupOperation,
+	type YdbAlterTableAction,
+	type YdbAlterTableOperation,
+	type YdbAlterTopicAction,
+	type YdbAlterTopicOperation,
+	type YdbAlterTransferOperation,
+	type YdbAlterTransferOptions,
+	type YdbAlterUserOperation,
+	type YdbAnalyzeOperation,
+	type YdbAsyncReplicationConsistencyLevel,
+	type YdbAsyncReplicationOptions,
+	type YdbAsyncReplicationTarget,
+	type YdbChangefeedFormat,
+	type YdbChangefeedMode,
+	type YdbChangefeedOptions,
+	type YdbCreateAsyncReplicationOperation,
+	type YdbCreateGroupOperation,
+	type YdbCreateSecretOperation,
+	type YdbCreateTopicOperation,
+	type YdbCreateTopicOptions,
+	type YdbCreateTransferOperation,
+	type YdbCreateUserOperation,
+	type YdbCreateViewOperation,
+	type YdbCreateViewOptions,
+	type YdbDropAsyncReplicationOperation,
+	type YdbDropChangefeedOperation,
+	type YdbDropGroupOperation,
+	type YdbDropTopicOperation,
+	type YdbDropTransferOperation,
+	type YdbDropUserOperation,
+	type YdbDropViewOperation,
+	type YdbGrantOperation,
+	type YdbInlineMigration,
+	type YdbMigrationLockConfig,
+	type YdbMigrationOperation,
+	type YdbMigrationRecoveryConfig,
+	type YdbMigrationStatus,
+	type YdbRenameTableOperation,
+	type YdbRevokeOperation,
+	type YdbShowCreateObjectType,
+	type YdbShowCreateOperation,
+	type YdbTopicConsumer,
+	type YdbTransferOptions,
+	type YdbUserOptions,
+} from './ydb/migration-ddl.js'
