+++ third-parties/drizzle-adapter/src/sql.ts
@@ -1,0 +1,88 @@
+// YQL expression helpers: built-in functions, UDF wrappers, set operators,
+// grouping/windowing primitives, hints, pragmas, scripts. These all return
+// `SQL` fragments that compose with drizzle's query builders.
+
+export {
+	currentUtcDate,
+	currentUtcDatetime,
+	currentUtcTimestamp,
+	maxOf,
+	minOf,
+	random,
+	randomNumber,
+	randomUuid,
+	unwrap,
+} from './ydb-core/query-builders/builtins.js'
+export { crc32c, crc64, numericHash, xxHash } from './ydb-core/query-builders/digest.js'
+export {
+	asTable,
+	cube,
+	distinctHint,
+	groupKey,
+	grouping,
+	groupingSets,
+	hop,
+	hopEnd,
+	hopStart,
+	knnCosineDistance,
+	knnCosineSimilarity,
+	knnDistance,
+	knnEuclideanDistance,
+	knnInnerProductSimilarity,
+	knnManhattanDistance,
+	knnSimilarity,
+	matchRecognize,
+	rollup,
+	sessionStart,
+	sessionWindow,
+	uniqueHint,
+	values,
+	valuesTable,
+	windowDefinition,
+	type YdbFlattenConfig,
+	type YdbFlattenMode,
+	type YdbGroupingSet,
+	type YdbKnnDistanceFunction,
+	type YdbKnnSimilarityFunction,
+	type YdbMatchRecognizeConfig,
+	type YdbSampleConfig,
+	type YdbUniqueDistinctHint,
+	type YdbValuesOptions,
+	type YdbValuesPrimitive,
+	type YdbValuesRow,
+	type YdbWindowClause,
+	type YdbWindowDefinitionConfig,
+} from './ydb-core/query-builders/select-syntax.js'
+export {
+	except,
+	intersect,
+	union,
+	unionAll,
+} from './ydb-core/query-builders/select.set-operators.js'
+export {
+	commit,
+	declareParam,
+	defineAction,
+	doAction,
+	doBlock,
+	intoResult,
+	kMeansTreeSearchTopSize,
+	pragma,
+	yqlScript,
+	type YdbActionParameter,
+	type YdbScriptExpression,
+	type YdbScriptPrimitive,
+} from './ydb-core/query-builders/yql-script.js'
+export type {
+	YdbDeleteConfig,
+	YdbFlatRelationalQueryConfig,
+	YdbInsertConfig,
+	YdbJoinConfig,
+	YdbJoinType,
+	YdbRelationalQueryConfig,
+	YdbRelationalQueryResult,
+	YdbSelectConfig,
+	YdbSetOperatorConfig,
+	YdbSetOperatorSource,
+	YdbUpdateConfig,
+} from './ydb/dialect.js'
