+++ third-parties/drizzle-adapter/src/ydb-core/table.test.ts
@@ -2,7 +2,7 @@ import { test } from 'vitest'
 import * as assert from 'node:assert/strict'
 import { Table, getTableName } from 'drizzle-orm/table'
 import { sql as yql } from 'drizzle-orm'
-import { customType, integer, text, uuid, ydbTable, ydbTableCreator } from '../index.ts'
+import { customType, integer, text, uuid, ydbTable, ydbTableCreator } from '../schema.ts'
 import { getYdbColumnBuilders, ydbColumnBuilders } from '../ydb-core/columns/all.ts'
 import { YdbColumn } from '../ydb-core/columns/common.ts'
 import { getTableConfig } from '../ydb-core/table.utils.ts'
