+++ third-parties/drizzle-adapter/src/ydb/createDrizzle.transaction.test.ts
@@ -2,7 +2,8 @@ import { test } from 'vitest'
 import * as assert from 'node:assert/strict'
 import { relations } from 'drizzle-orm'
 import { TransactionRollbackError } from 'drizzle-orm/errors'
-import { type YdbExecuteOptions, drizzle, integer, text, ydbTable } from '../index.ts'
+import { type YdbExecuteOptions, drizzle } from '../index.ts'
+import { integer, text, ydbTable } from '../schema.ts'
 
 let users = ydbTable('users', {
 	id: integer('id').notNull(),
