+++ third-parties/drizzle-adapter/src/ydb/dialect.methods.test.ts
@@ -4,7 +4,7 @@ import { aliasedTable } from 'drizzle-orm/alias'
 import { createTableRelationsHelpers, extractTablesRelationalConfig } from 'drizzle-orm/relations'
 import { eq, sql as yql } from 'drizzle-orm'
 import { WithSubquery } from 'drizzle-orm/subquery'
-import { date, datetime, decimal, json, text, timestamp, uuid, ydbTable } from '../index.ts'
+import { date, datetime, decimal, json, text, timestamp, uuid, ydbTable } from '../schema.ts'
 import { YdbDialect } from '../ydb/dialect.ts'
 import { YdbSelectBuilder } from '../ydb-core/query-builders/index.ts'
 import { orderSelectedFields } from '../ydb-core/result-mapping.ts'
